CostValueCard

+++ Assets/Scrpits/Card/CardData/CardData.cs
@@ -8,7 +8,7 @@ public class CardData : ScriptableObject
 {
     public int HP;
     public int Damage;
-    public float Cost;
+    public int Cost;
     public string cardInfo;
     public void ApplyDamage(int damage)
     {
